Create README.md
This batch script was made with ChatGPT! The goal is to disable un-necessary services, Clear temp, dump and cache files automatically, And check for and install windows updates. 
The goal of this script is to automate some basic maintenence tasks for your PC, And if you have minimal free space on your C drive you should see an improvement in performance!

You can help improve this script by creating an issue and suggesting programs that take up a lot of cache space, Or by fixing any errors you see.

There are a few errors, as my knowledge on scripting and coding is very minimal, Hence why i built this using ChatGPT. 
With that being said, Any errors you see will simply be due to a lack of permission to access a certain file, certain files being used, or a pathing issue. 
the windows update script is currently broken which i would really like to get working.